extract parts of get_metric.py script to fl_prog package
